requirements.txt
lxml
python-pptx
python-docx
openpyxl
boto3
